+++ LumbApp/Expertos/ExpertoZE/InformeZE.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace LumbApp.Expertos.ExpertoZE
+﻿namespace LumbApp.Expertos.ExpertoZE
 {
     public class InformeZE
     {
